[Refactor] P2P split Phase 1: move P2P tests into tests/p2p
Pure relocation, zero assertion change (Phase 1 commit 4/4):
- git mv all P2P unit/integration tests (incl. tests/ha/oplog/*) into
  mooncake-store/tests/p2p/, per §4.G of the split plan
- add tests/p2p/CMakeLists.txt with add_p2p_store_test (same recipe as
  add_store_test); redis-conditional registrations preserved
- adjust only include paths (utils/common.h -> ../utils/common.h)
- stress_single_workload_runner.py moves with the P2P test assets
- e2e/ moves are deferred to Phase 4 (shared harness + binary switch)
