+++ mooncake-store/tests/CMakeLists.txt
@@ -22,7 +22,6 @@ add_store_test(master_service_ssd_test master_service_ssd_test.cpp)
 add_store_test(client_integration_test client_integration_test.cpp)
 add_store_test(cxl_client_integration_test cxl_client_integration_test.cpp)
 add_store_test(master_metrics_test master_metrics_test.cpp)
-add_store_test(p2p_master_metrics_test p2p_master_metrics_test.cpp)
 add_store_test(posix_file_test posix_file_test.cpp)
 add_store_test(thread_pool_test thread_pool_test.cpp)
 add_store_test(transfer_task_test transfer_task_test.cpp)
@@ -31,69 +30,21 @@ add_store_test(offset_allocator_test offset_allocator_test.cpp)
 add_store_test(utils_test utils_test.cpp)
 add_store_test(client_buffer_test client_buffer_test.cpp)
 add_store_test(client_config_builder_test client_config_builder_test.cpp)
-add_store_test(runtime_config_store_test runtime_config_store_test.cpp)
 add_store_test(pybind_client_test pybind_client_test.cpp)
 add_store_test(ipv6_client_test ipv6_client_test.cpp)
 add_store_test(client_metrics_test client_metrics_test.cpp)
-add_store_test(client_metrics_aggregator_test client_metrics_aggregator_test.cpp)
-add_store_test(client_http_metrics_test client_http_metrics_test.cpp)
 add_store_test(serializer_test serializer_test.cpp)
 add_store_test(non_ha_reconnect_test non_ha_reconnect_test.cpp)
 add_store_test(heartbeat_dedicated_port_test heartbeat_dedicated_port_test.cpp)
 add_store_test(real_client_remount_test real_client_remount_test.cpp)
 add_store_test(shm_helper_test shm_helper_test.cpp)
 add_store_test(storage_backend_test storage_backend_test.cpp)
-add_store_test(storage_tier_test storage_tier_test.cpp)
 add_store_test(mutex_test mutex_test.cpp)
 add_store_test(file_storage_test file_storage_test.cpp)
-add_store_test(tiered_backend_test tiered_backend_test.cpp)
-if(USE_ASCEND_CACHE_TIER)
-    add_store_test(ascend_tier_test ascend_tier_test.cpp)
-endif()
-add_store_test(data_manager_test data_manager_test.cpp)
-add_store_test(inflight_tracker_test inflight_tracker_test.cpp)
-add_store_test(client_rpc_service_test client_rpc_service_test.cpp)
-add_store_test(peer_client_test peer_client_test.cpp)
 add_store_test(centralized_client_meta_test centralized_client_meta_test.cpp)
-add_store_test(p2p_master_service_test p2p_master_service_test.cpp)
-add_store_test(p2p_master_http_endpoints_test p2p_master_http_endpoints_test.cpp)
-add_store_test(p2p_client_ha_test p2p_client_ha_test.cpp)
-add_store_test(p2p_client_integration_test p2p_client_integration_test.cpp)
-add_store_test(p2p_client_http_endpoints_test p2p_client_http_endpoints_test.cpp)
-add_store_test(p2p_register_concurrency_test p2p_register_concurrency_test.cpp)
-add_store_test(peer_client_perf_test peer_client_perf_test.cpp)
-add_store_test(scheduler_integration_test scheduler_integration_test.cpp)
-add_store_test(tier_role_resolver_test tier_role_resolver_test.cpp)
-add_store_test(tinylfu_sketch_test tinylfu_sketch_test.cpp)
-add_store_test(multi_lru_test multi_lru_test.cpp)
-add_store_test(bounded_dedup_queue_test bounded_dedup_queue_test.cpp)
-add_store_test(event_driven_policy_test event_driven_policy_test.cpp)
-add_store_test(event_driven_scheduler_test event_driven_scheduler_test.cpp)
-add_store_test(p2p_segment_manager_test p2p_segment_manager_test.cpp)
-add_store_test(p2p_client_meta_test p2p_client_meta_test.cpp)
-add_store_test(p2p_client_manager_test p2p_client_manager_test.cpp)
 add_store_test(centralized_client_manager_test centralized_client_manager_test.cpp)
-add_store_test(route_cache_test route_cache_test.cpp)
-add_store_test(async_metadata_notifier_test async_metadata_notifier_test.cpp)
-add_store_test(ha_recovery_manager_test ha_recovery_manager_test.cpp)
-add_store_test(ha_integration_test ha_integration_test.cpp)
 add_store_test(task_manager_test task_manager_test.cpp)
-
-# HA OpLog ported tests
-add_store_test(oplog_manager_test ha/oplog/oplog_manager_test.cpp)
-add_store_test(oplog_serializer_test ha/oplog/oplog_serializer_test.cpp)
-add_store_test(localfs_oplog_store_test ha/oplog/localfs_oplog_store_test.cpp)
-add_store_test(standby_state_machine_test ha/oplog/standby_state_machine_test.cpp)
-add_store_test(ha_metric_manager_test ha/oplog/ha_metric_manager_test.cpp)
-add_store_test(oplog_applier_test ha/oplog/oplog_applier_test.cpp)
-add_store_test(oplog_replicator_test ha/oplog/oplog_replicator_test.cpp)
-add_store_test(p2p_oplog_test
-    ha/oplog/p2p_oplog_test.cpp
-    ha/oplog/p2p_oplog_applier_test.cpp
-    ha/oplog/p2p_record_oplog_test.cpp
-    ha/oplog/p2p_standby_metadata_store_test.cpp
-    ha/oplog/p2p_hot_standby_service_test.cpp
-)
+add_subdirectory(p2p)
 add_subdirectory(e2e)
 
 add_executable(high_availability_test high_availability_test.cpp)
@@ -102,15 +53,6 @@ if (STORE_USE_ETCD)
     add_test(NAME high_availability_test COMMAND high_availability_test)
 endif()
 
-add_executable(redis_election_helper_test redis_election_helper_test.cpp)
-target_link_libraries(redis_election_helper_test PUBLIC ${MOONCAKE_STORE_LIBS} transfer_engine cachelib_memory_allocator ${ETCD_WRAPPER_LIB} glog gtest gtest_main pthread)
-if(STORE_USE_REDIS)
-    add_test(NAME redis_election_helper_test COMMAND redis_election_helper_test)
-endif()
-
-if(STORE_USE_REDIS)
-    add_store_test(redis_oplog_store_test ha/oplog/redis_oplog_store_test.cpp)
-endif()
 
 add_executable(stress_workload_test stress_workload_test.cpp)
 target_link_libraries(stress_workload_test PUBLIC
