+++ mooncake-store/tests/heartbeat_dedicated_port_test.cpp
@@ -47,7 +47,7 @@
 #include "master_config.h"
 #include "p2p/master/p2p_master_client.h"
 #include "rpc_types.h"
-#include "test_p2p_server_helpers.h"
+#include "p2p/test_p2p_server_helpers.h"
 #include "test_server_helpers.h"
 #include "types.h"
 #include "utils.h"
