+++ mooncake-store/tests/p2p/client_rpc_service_test.cpp
@@ -10,7 +10,7 @@
 #include "p2p/client/client_rpc_types.h"
 #include "p2p/client/data_manager.h"
 #include "p2p/tiered_cache/tiered_backend.h"
-#include "utils/common.h"
+#include "../utils/common.h"
 #include "transfer_engine.h"
 #include "types.h"
