+++ mooncake-store/tests/p2p/data_manager_test.cpp
@@ -14,7 +14,7 @@
 #define protected public
 #include "p2p/client/data_manager.h"
 #include "p2p/tiered_cache/tiered_backend.h"
-#include "utils/common.h"
+#include "../utils/common.h"
 #include "transfer_engine.h"
 #include "types.h"
 #include "utils.h"
