+++ mooncake-store/tests/p2p/event_driven_scheduler_test.cpp
@@ -13,7 +13,7 @@
 #include "p2p/client/p2p_client_metric.h"
 #include "p2p/tiered_cache/tiered_backend.h"
 #include "p2p/tiered_cache/tiers/cache_tier.h"  // TempDRAMBuffer
-#include "utils/common.h"                   // InitTieredBackendForTest
+#include "../utils/common.h"                   // InitTieredBackendForTest
 
 namespace mooncake {
 namespace {
