+++ mooncake-store/tests/p2p/ha/oplog/mock_metadata_store.h
