+++ mooncake-store/tests/p2p/ha/oplog/mock_oplog_store.h
