+++ mooncake-store/tests/p2p/peer_client_perf_test.cpp
@@ -22,7 +22,7 @@
 #include "p2p/client/data_manager.h"
 #include "p2p/client/peer_client.h"
 #include "p2p/tiered_cache/tiered_backend.h"
-#include "utils/common.h"
+#include "../utils/common.h"
 #include "transfer_engine.h"
 #include "types.h"
 #include "utils.h"
