+++ mooncake-store/tests/p2p/peer_client_test.cpp
@@ -19,7 +19,7 @@
 #include "p2p/client/data_manager.h"
 #include "p2p/client/peer_client.h"
 #include "p2p/tiered_cache/tiered_backend.h"
-#include "utils/common.h"
+#include "../utils/common.h"
 #include "transfer_engine.h"
 #include "types.h"
