+++ mooncake-store/tests/p2p/tiered_backend_test.cpp
@@ -15,7 +15,7 @@
 #include "p2p/client/p2p_client_metric.h"
 #include "p2p/tiered_cache/tiered_backend.h"
 #include "p2p/tiered_cache/tiers/cache_tier.h"
-#include "utils/common.h"
+#include "../utils/common.h"
 
 // Helper function to parse JSON string using thread-safe CharReaderBuilder
 static bool parseJsonString(const std::string& json_str, Json::Value& value,
